pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.109.1",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
]
